Crate gsys

Declares gear protocol syscalls for WASM.

Structs

BlockNumberWithHash

Represents type defining concatenated block number with hash. 36 bytes.

EnvVars

Current version of execution settings.

ErrorBytes

ErrorWithGas

Represents type defining concatenated error code with gas. 12 bytes.

ErrorWithHandle

Represents type defining concatenated length with handle. 8 bytes.

ErrorWithHash

Represents type defining concatenated hash with error code. 36 bytes.

ErrorWithReplyCode

Represents type defining concatenated reply code with error code. 8 bytes.

ErrorWithSignalCode

Represents type defining concatenated signal code with length. 8 bytes.

ErrorWithTwoHashes

Represents type defining concatenated two hashes with error code. 68 bytes.

GasMultiplier

Type representing converter between gas and value.

HashWithValue

Represents type defining concatenated hash with value. 48 bytes.

Percent

Basic struct for working with integer percentages allowing values greater than 100.

TwoHashes

Represents type defining concatenated two hashes. 64 bytes.

TwoHashesWithValue

Represents type defining concatenated two hashes with value. 80 bytes.

Functions

gr_block_height⚠

Infallible gr_block_height get syscall.

gr_block_timestamp⚠

Infallible gr_block_timestamp get syscall.

gr_create_program⚠

Fallible gr_create_program send syscall.

gr_create_program_wgas⚠

Fallible gr_create_program_wgas send syscall.

gr_debug⚠

Infallible gr_debug info syscall.

gr_env_vars⚠

Infallible gr_env_vars get syscall. It leaves backend with unrecoverable error if incorrect version is passed.

gr_exit⚠

Infallible gr_exit control syscall.

gr_gas_available⚠

Infallible gr_gas_available get syscall.

gr_leave⚠

Infallible gr_leave control syscall.

gr_message_id⚠

Infallible gr_message_id get syscall.

gr_oom_panic⚠

Infallible gr_oom_panic control syscall.

gr_panic⚠

Infallible gr_panic control syscall.

gr_program_id⚠

Infallible gr_program_id get syscall.

gr_random⚠

Infallible gr_random calculate syscall.

gr_read⚠

Fallible gr_read get syscall.

gr_reply⚠

Fallible gr_reply send syscall.

gr_reply_code⚠

Fallible gr_reply_code get syscall.

gr_reply_commit⚠

Fallible gr_reply_commit send syscall.

gr_reply_commit_wgas⚠

Fallible gr_reply_commit_wgas send syscall.

gr_reply_deposit⚠

Fallible gr_reply_deposit control syscall.

gr_reply_input⚠

Fallible gr_reply_input send syscall.

gr_reply_input_wgas⚠

Fallible gr_reply_input_wgas send syscall.

gr_reply_push⚠

Fallible gr_reply_push send syscall.

gr_reply_push_input⚠

Fallible gr_reply_push_input send syscall.

gr_reply_to⚠

Fallible gr_reply_to get syscall.

gr_reply_wgas⚠

Fallible gr_reply_wgas send syscall.

gr_reservation_reply⚠

Fallible gr_reservation_reply send syscall.

gr_reservation_reply_commit⚠

Fallible gr_reservation_reply_commit send syscall.

gr_reservation_send⚠

Fallible gr_reservation_send send syscall.

gr_reservation_send_commit⚠

Fallible gr_reservation_send_commit send syscall.

gr_reserve_gas⚠

Fallible gr_reserve_gas control syscall.

gr_send⚠

Fallible gr_send send syscall.

gr_send_commit⚠

Fallible gr_send_commit send syscall.

gr_send_commit_wgas⚠

Fallible gr_send_commit_wgas send syscall.

gr_send_init⚠

Fallible gr_send_init send syscall.

gr_send_input⚠

Fallible gr_send_input send syscall.

gr_send_input_wgas⚠

Fallible gr_send_input_wgas send syscall.

gr_send_push⚠

Fallible gr_send_push send syscall.

gr_send_push_input⚠

Fallible gr_send_push_input send syscall.

gr_send_wgas⚠

Fallible gr_send_wgas send syscall.

gr_signal_code⚠

Fallible gr_signal_code get syscall.

gr_signal_from⚠

Fallible gr_signal_from get syscall.

gr_size⚠

Infallible gr_size get syscall.

gr_source⚠

Infallible gr_source get syscall.

gr_system_reserve_gas⚠

Fallible gr_system_reserve_gas control syscall.

gr_unreserve_gas⚠

Fallible gr_unreserve_gas control syscall.

gr_value⚠

Infallible gr_value get syscall.

gr_value_available⚠

Infallible gr_value_available get syscall.

gr_wait⚠

Infallible gr_wait control syscall.

gr_wait_for⚠

Infallible gr_wait_for control syscall.

gr_wait_up_to⚠

Infallible gr_wait_up_to control syscall.

gr_wake⚠

Fallible gr_wake control syscall.

Type Aliases

BlockCount

Represents block count type.

BlockNumber

Represents block number type.

BlockTimestamp

Represents block number type.

BufferStart

Represents byte type, which is a start of a buffer.

ErrorCode

Represents error code type.

Gas

Represents gas type.

Handle

Represents handle type.

Hash

Represents hash type.

Length

Represents length type.

Offset

Represents offset type.

ReplyCode

Represents reply code type.

SignalCode

Represents signal code type.

SizedBufferStart

Represents byte type, which is a start of a sized buffer.

Value

Represents value type.